debug stuff

+++ include/float_types.h
@@ -1,7 +1,7 @@
 #ifndef NDG_FLOAT_TYPES_H
 #define NDG_FLOAT_TYPES_H
 
-typedef float hostFloat;
-typedef float deviceFloat;
+typedef double hostFloat;
+typedef double deviceFloat;
 
 #endif
